# openapi.yaml
openapi: 3.0.3
info:
  title: Card Game API
  version: 1.0.0
  description: API specification for Auth, Player, and Game services.

servers:
  - url: http://localhost:8001
    description: Auth Service
  - url: http://localhost:8002
    description: Player Service
  - url: http://localhost:8003
    description: Game Service

paths:

  ###############################################
  # AUTH SERVICE (PORT 8001)
  ###############################################

  /auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User registered

  /auth/login:
    post:
      summary: Login and obtain JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: JWT token returned

  /auth/validate:
    get:
      summary: Validate JWT token
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token is valid

  /auth/health:
    get:
      summary: Health check
      tags: [Auth]
      responses:
        "200":
          description: Alive

  ###############################################
  # PLAYER SERVICE (PORT 8002)
  ###############################################

  /health:
    get:
      summary: Health check
      tags: [Player]
      responses:
        "200":
          description: Service OK

  /cards:
    get:
      summary: List all cards (public)
      tags: [Player]
      responses:
        "200":
          description: A list of cards

  /cards/{card_id}:
    get:
      summary: Get card details (public)
      tags: [Player]
      parameters:
        - name: card_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card details

  /players:
    post:
      summary: Create or update my profile
      tags: [Player]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Profile updated
        "201":
          description: Profile created

  /players/me:
    get:
      summary: Get my profile
      tags: [Player]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
        "404":
          description: Player not found

  /matches:
    post:
      summary: Create match record (for statistics)
      tags: [Player]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - player1_external_id
                - player2_external_id
                - player1_score
                - player2_score
                - external_match_id
              properties:
                player1_external_id:
                  type: string
                player2_external_id:
                  type: string
                winner_external_id:
                  type: string
                  nullable: true
                player1_score:
                  type: integer
                player2_score:
                  type: integer
                external_match_id:
                  type: string
                  description: Idempotency key.
                turns:
                  type: array
                  items:
                    type: object
                    properties:
                      turn_number:
                        type: integer
                      player1_card_id:
                        type: integer
                      player2_card_id:
                        type: integer
                      winner_external_id:
                        type: string
                        nullable: true
      responses:
        "201":
          description: Match record created

  /players/{player_id}/matches:
    get:
      summary: Player match history (public)
      tags: [Player]
      parameters:
        - name: player_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of matches

  /players/{player_id}/matches/{match_id}:
    get:
      summary: Get match details (public)
      tags: [Player]
      parameters:
        - name: player_id
          in: path
          required: true
          schema:
            type: string
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match details

  /leaderboard:
    get:
      summary: Global leaderboard (public)
      tags: [Player]
      responses:
        "200":
          description: A list of player rankings

  ###############################################
  # GAME SERVICE (PORT 8003)
  # API Gateway Prefix: /game/*
  ###############################################

  /game/matches:
    post:
      summary: Create new match (protected)
      tags: [Game]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - player1_id
                - player2_id
              properties:
                player1_id:
                  type: string
                  pattern: "^[a-zA-Z0-9_-]+$"
                  minLength: 1
                  maxLength: 100
                  description: "Username (alphanumeric, underscore, hyphen)"
                  example: "alice"
                player2_id:
                  type: string
                  pattern: "^[a-zA-Z0-9_-]+$"
                  minLength: 1
                  maxLength: 100
                  description: "Opponent username"
                  example: "bob"
      responses:
        "200":
          description: Match created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_id:
                    type: string
                  player_id:
                    type: string
                  hand:
                    type: array
                    items:
                      type: object
                  status:
                    type: string
        "403":
          description: Forbidden - can only create matches as yourself
        "422":
          description: Validation error

  /game/matches/{match_id}/move:
    post:
      summary: Submit a card move (protected)
      tags: [Game]
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
          description: Match identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - player_id
                - match_card_id
              properties:
                player_id:
                  type: string
                  pattern: "^[a-zA-Z0-9_-]+$"
                  description: "Player making the move"
                match_card_id:
                  type: integer
                  minimum: 1
                  maximum: 100000
                  description: "ID of card to play from hand"
      responses:
        "200":
          description: Move accepted
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: "Waiting for opponent"
                    properties:
                      status:
                        type: string
                        enum: ["waiting_for_opponent"]
                  - type: object
                    description: "Round result"
                    properties:
                      round:
                        type: integer
                      winner:
                        type: string
                        nullable: true
                      reason:
                        type: string
                      points_p1:
                        type: integer
                      points_p2:
                        type: integer
                      match_finished:
                        type: boolean
                      match_winner:
                        type: string
                        nullable: true
        "400":
          description: Invalid move
        "403":
          description: Forbidden

  /game/matches/{match_id}:
    get:
      summary: Get match state (protected)
      tags: [Game]
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
          description: Match identifier
        - name: player_id
          in: query
          required: true
          schema:
            type: string
          description: Player identifier (must match authenticated user)
      responses:
        "200":
          description: Match state retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_id:
                    type: string
                  status:
                    type: string
                  current_round:
                    type: integer
                  points_p1:
                    type: integer
                  points_p2:
                    type: integer
                  player_hand:
                    type: array
                    items:
                      type: object
                  used_cards:
                    type: array
                    items:
                      type: object
                  opponent_used_cards:
                    type: array
                    items:
                      type: object
                  match_winner:
                    type: string
                    nullable: true
        "403":
          description: Forbidden
        "404":
          description: Match not found

  /game/cards:
    get:
      summary: Get deck gallery as SVG (public)
      tags: [Game]
      description: "Displays all available cards in organized grid (5 per row) with statistics"
      responses:
        "200":
          description: SVG deck gallery
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary

  /game/cards/{card_id}:
    get:
      summary: Get card detail as SVG (public)
      tags: [Game]
      description: "Shows single card with category, power level, and rarity"
      parameters:
        - name: card_id
          in: path
          required: true
          schema:
            type: integer
          description: Card definition ID
      responses:
        "200":
          description: SVG card detail
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        "404":
          description: Card not found

  /game/matches/{match_id}/hand:
    get:
      summary: Get player hand visualization as SVG (protected)
      tags: [Game]
      security:
        - bearerAuth: []
      description: "Shows player's cards with visual distinction between available and used cards"
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
          description: Match identifier
        - name: player_id
          in: query
          required: true
          schema:
            type: string
          description: Player identifier (must match authenticated user)
      responses:
        "200":
          description: SVG player hand visualization
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
        "404":
          description: Match not found

  /game/health:
    get:
      summary: Health check
      tags: [Game]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT